# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My API",
    "description": "This is a sample Spring Boot 3.5 API with Swagger",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9001",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/comments/{id}": {
      "put": {
        "tags": [
          "Comments"
        ],
        "summary": "Update a comment",
        "description": "Updates an existing comment by ID. Only the comment owner or admin can update a comment.",
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the comment to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "404": {
            "description": "Comment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Comments"
        ],
        "summary": "Delete a comment",
        "description": "Deletes a comment by ID. Only the comment owner or admin can delete a comment.",
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the comment to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted successfully"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      }
    },
...
    "/api/comments": {
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Create a new comment",
        "description": "Creates a new comment on a blog post. Can be either a top-level comment or a reply to an existing comment.",
        "operationId": "createComment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments/post/{postId}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get paginated comments for a post",
        "description": "Retrieves a paginated list of comments for a specific blog post.",
        "operationId": "getCommentsForPost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of the post to retrieve comments for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination and sorting parameters",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved comments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid post ID",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCommentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments/post/{postId}/tree": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get comment tree for a post",
        "description": "Retrieves all comments for a post in a hierarchical tree structure with replies nested under their parent comments.",
        "operationId": "getCommentTreeForPost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of the post to retrieve comment tree for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved comment tree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments/post/{postId}/count": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get comment count for a post",
        "description": "Returns the total number of comments for a specific blog post.",
        "operationId": "getCommentCount",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of the post to get comment count for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved comment count",
            "content": {
              "text/plain": {

              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentDto": {
        "type": "object",
        "description": "Data Transfer Object for Comment operations",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string",
            "description": "The content of the comment",
            "example": "This is a great post! Thanks for sharing."
          },
          "authorId": {
            "type": "string",
            "description": "ID of the comment author",
            "example": "auth0|1234567890"
          },
          "authorName": {
            "type": "string",
            "description": "Name of the comment author",
            "example": "John Doe"
          },
          "authorEmail": {
            "type": "string",
            "description": "Email of the comment author",
            "example": "john.doe@example.com"
          },
          "approved": {
            "type": "boolean",
            "description": "Whether the comment has been approved by a moderator",
            "example": false
          },
          "postId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the post this comment belongs to",
            "example": 123
          },
          "parentId": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the parent comment if this is a reply",
            "example": 456
          },
          "replies": {
            "type": "array",
            "description": "List of replies to this comment",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "authorId",
          "content",
          "postId"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageCommentDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      }
    }
  }
}